pyproject.toml
[tool.poetry]
name = "with-langchain"
version = "0.1.0"
description = ""
authors = ["sam <sam.barrowclough@gmail.com>"]
readme = "README.md"
packages = [{include = "with_langchain"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.172"
openai = "^0.27.6"
pandas = "^2.0.1"
tabulate = "^0.9.0"
matplotlib = "^3.7.1"
python-dotenv = "^1.0.0"
supabase = "^1.0.3"
lanarky = "^0.6.2"
uvicorn = "^0.22.0"
flask = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
